requirements.txt
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.20.0
scikit-learn>=1.3.0
transformers==4.44.2
tensorflow>=2.17.0
tf-keras>=2.17.0
ml_dtypes>=0.5.0
huggingface_hub>=0.23.0
pillow>=9.5.0
requests>=2.31.0
pytest>=7.3.0
torch>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0